update: custom hook on main page to handle api call
